Refactor: Refactor LoginModal component
* Unify LoginBoxBody component to LoginModal

* Modify class names of elements

* Modify render logic of LoginModal component
